# Servico/API Pizzaria/pizzaria.api01/pizzaria.api01/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pizzaria.api01, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/MateriaPrima/{id}": {
      "get": {
        "tags": [
          "MateriaPrima"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MateriaPrima/Listar": {
      "get": {
        "tags": [
          "MateriaPrima"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MateriaPrima/Incluir": {
      "post": {
        "tags": [
          "MateriaPrima"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MateriaPrimas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MateriaPrima/Alterar/{id}": {
      "put": {
        "tags": [
          "MateriaPrima"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MateriaPrimas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MateriaPrima/Excluir/{id}": {
      "delete": {
        "tags": [
          "MateriaPrima"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/Listar": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/Incluir": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Produtos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/Alterar/{id}": {
      "put": {
        "tags": [
          "Produtos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Produtos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/Excluir/{id}": {
      "delete": {
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Produtos/ListarMateriaPrima": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Listar": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Incluir": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Alterar/{id}": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Excluir/{id}": {
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "MateriaPrimas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descricao": {
          "type": "string"
        },
        "estoque": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Produtos": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "codigo": {
          "format": "int32",
          "type": "integer"
        },
        "descricao": {
          "type": "string"
        },
        "quantidade": {
          "format": "int32",
          "type": "integer"
        },
        "materiaPrimaId": {
          "format": "int32",
          "type": "integer"
        },
        "descricaoMateriaPrima": {
          "type": "string"
        }
      }
    },
    "Usuarios": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        }
      }
    }
  }
}